Fixed minor issues with encoding to float

+++ genetic_algorithms_py/encoder.py
@@ -8,4 +8,4 @@ def e(value):
 
 def d(value):
     decoded_integer = BitArray(bin=value).int
-    return float(decoded_integer / multiplier)
+    return long(decoded_integer / multiplier)
